+++ 2. java/Ejercicios Integradores/P2VIVO/src/ejercicio2/clases/Moto.java
@@ -1,7 +1,7 @@
 package ejercicio2.clases;
 
 public class Moto extends Vehiculo{
-    public Moto(float velocidad, float aceleracion, float anguloDeGiro, String patente) {
+    public Moto(double velocidad, double aceleracion, double anguloDeGiro, String patente) {
         super(velocidad, aceleracion, anguloDeGiro, patente, 300, 2);
     }
