+++ 2. java/Ejercicios Integradores/P2VIVO/src/ejercicio2/clases/SocorristaAuto.java
@@ -2,7 +2,7 @@ package ejercicio2.clases;
 
 public class SocorristaAuto extends Auto{
 
-    public SocorristaAuto(float velocidad, float aceleracion, float anguloDeGiro, String patente) {
+    public SocorristaAuto(double velocidad, double aceleracion, double anguloDeGiro, String patente) {
         super(velocidad, aceleracion, anguloDeGiro, patente);
     }
