+++ 2. java/Ejercicios Integradores/P2VIVO/src/ejercicio2/clases/SocorristaMoto.java
@@ -1,7 +1,7 @@
 package ejercicio2.clases;
 
 public class SocorristaMoto extends Moto{
-    public SocorristaMoto(float velocidad, float aceleracion, float anguloDeGiro, String patente) {
+    public SocorristaMoto(double velocidad, double aceleracion, double anguloDeGiro, String patente) {
         super(velocidad, aceleracion, anguloDeGiro, patente);
     }
